Put every output module into a separate file
