+++ src/data.c
@@ -24,54 +24,15 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdbool.h>
-#include <limits.h>
-// gethostname() needs _XOPEN_SOURCE 500 on unistd.h
-#ifndef _XOPEN_SOURCE
-#define _XOPEN_SOURCE 500
-#endif
 
 #ifndef _MSC_VER
 #include <unistd.h>
 #endif
 
-#ifdef _WIN32
-    #if !defined(_WIN32_WINNT) || (_WIN32_WINNT < 0x0600)
-    #undef _WIN32_WINNT
-    #define _WIN32_WINNT 0x0600   /* Needed to pull in 'struct sockaddr_storage' */
-    #endif
-
-    #include <winsock2.h>
-    #include <ws2tcpip.h>
-#else
-    #include <sys/types.h>
-    #include <sys/socket.h>
-    #include <netdb.h>
-    #include <netinet/in.h>
-
-    #define SOCKET          int
-    #define INVALID_SOCKET  (-1)
-    #define closesocket(x)  close(x)
-#endif
-
-#include <time.h>
-
-#include "term_ctl.h"
 #include "abuf.h"
 #include "fatal.h"
 
 #include "data.h"
-
-#ifdef _WIN32
-    #define _POSIX_HOST_NAME_MAX  128
-    #define perror(str)           ws2_perror(str)
-
-    static void ws2_perror (const char *str)
-    {
-        if (str && *str)
-            fprintf(stderr, "%s: ", str);
-        fprintf(stderr, "Winsock error %d.\n", WSAGetLastError());
-    }
-#endif
 
 typedef void* (*array_elementwise_import_fn)(void*);
 typedef void (*array_element_release_fn)(void*);
@@ -459,462 +420,6 @@ void print_array_value(data_output_t *output, data_array_t *array, char const *f
     }
 }
 
-/* JSON printer */
-
-static void print_json_array(data_output_t *output, data_array_t *array, char const *format)
-{
-    fprintf(output->file, "[");
-    for (int c = 0; c < array->num_values; ++c) {
-        if (c)
-            fprintf(output->file, ", ");
-        print_array_value(output, array, format, c);
-    }
-    fprintf(output->file, "]");
-}
-
-static void print_json_data(data_output_t *output, data_t *data, char const *format)
-{
-    bool separator = false;
-    fputc('{', output->file);
-    while (data) {
-        if (separator)
-            fprintf(output->file, ", ");
-        output->print_string(output, data->key, NULL);
-        fprintf(output->file, " : ");
-        print_value(output, data->type, data->value, data->format);
-        separator = true;
-        data = data->next;
-    }
-    fputc('}', output->file);
-}
-
-static void print_json_string(data_output_t *output, const char *str, char const *format)
-{
-    fprintf(output->file, "\"");
-    while (*str) {
-        if (*str == '"' || *str == '\\')
-            fputc('\\', output->file);
-        fputc(*str, output->file);
-        ++str;
-    }
-    fprintf(output->file, "\"");
-}
-
-static void print_json_double(data_output_t *output, double data, char const *format)
-{
-    fprintf(output->file, "%.3f", data);
-}
-
-static void print_json_int(data_output_t *output, int data, char const *format)
-{
-    fprintf(output->file, "%d", data);
-}
-
-static void data_output_json_free(data_output_t *output)
-{
-    if (!output)
-        return;
-
-    free(output);
-}
-
-struct data_output *data_output_json_create(FILE *file)
-{
-    data_output_t *output = calloc(1, sizeof(data_output_t));
-    if (!output) {
-        WARN_CALLOC("data_output_json_create()");
-        return NULL; // NOTE: returns NULL on alloc failure.
-    }
-
-    output->print_data   = print_json_data;
-    output->print_array  = print_json_array;
-    output->print_string = print_json_string;
-    output->print_double = print_json_double;
-    output->print_int    = print_json_int;
-    output->output_free  = data_output_json_free;
-    output->file         = file;
-
-    return output;
-}
-
-/* Pretty Key-Value printer */
-
-static int kv_color_for_key(char const *key)
-{
-    if (!key || !*key)
-        return TERM_COLOR_RESET;
-    if (!strcmp(key, "tag") || !strcmp(key, "time"))
-        return TERM_COLOR_BLUE;
-    if (!strcmp(key, "model") || !strcmp(key, "type") || !strcmp(key, "id"))
-        return TERM_COLOR_RED;
-    if (!strcmp(key, "mic"))
-        return TERM_COLOR_CYAN;
-    if (!strcmp(key, "mod") || !strcmp(key, "freq") || !strcmp(key, "freq1") || !strcmp(key, "freq2"))
-        return TERM_COLOR_MAGENTA;
-    if (!strcmp(key, "rssi") || !strcmp(key, "snr") || !strcmp(key, "noise"))
-        return TERM_COLOR_YELLOW;
-    return TERM_COLOR_GREEN;
-}
-
-static int kv_break_before_key(char const *key)
-{
-    if (!key || !*key)
-        return 0;
-    if (!strcmp(key, "model") || !strcmp(key, "mod") || !strcmp(key, "rssi") || !strcmp(key, "codes"))
-        return 1;
-    return 0;
-}
-
-static int kv_break_after_key(char const *key)
-{
-    if (!key || !*key)
-        return 0;
-    if (!strcmp(key, "id") || !strcmp(key, "mic"))
-        return 1;
-    return 0;
-}
-
-typedef struct {
-    struct data_output output;
-    void *term;
-    int color;
-    int ring_bell;
-    int term_width;
-    int data_recursion;
-    int column;
-} data_output_kv_t;
-
-#define KV_SEP "_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ "
-
-static void print_kv_data(data_output_t *output, data_t *data, char const *format)
-{
-    data_output_kv_t *kv = (data_output_kv_t *)output;
-
-    int color = kv->color;
-    int ring_bell = kv->ring_bell;
-
-    // top-level: update width and print separator
-    if (!kv->data_recursion) {
-        kv->term_width = term_get_columns(kv->term); // update current term width
-        if (color)
-            term_set_fg(kv->term, TERM_COLOR_BLACK);
-        if (ring_bell)
-            term_ring_bell(kv->term);
-        char sep[] = KV_SEP KV_SEP KV_SEP KV_SEP;
-        if (kv->term_width < (int)sizeof(sep))
-            sep[kv->term_width > 0 ? kv->term_width - 1 : 40] = '\0';
-        fprintf(output->file, "%s\n", sep);
-        if (color)
-            term_set_fg(kv->term, TERM_COLOR_RESET);
-    }
-    // nested data object: break before
-    else {
-        if (color)
-            term_set_fg(kv->term, TERM_COLOR_RESET);
-        fprintf(output->file, "\n");
-        kv->column = 0;
-    }
-
-    ++kv->data_recursion;
-    while (data) {
-        // break before some known keys
-        if (kv->column > 0 && kv_break_before_key(data->key)) {
-            fprintf(output->file, "\n");
-            kv->column = 0;
-        }
-        // break if not enough width left
-        else if (kv->column >= kv->term_width - 26) {
-            fprintf(output->file, "\n");
-            kv->column = 0;
-        }
-        // pad to next alignment if there is enough width left
-        else if (kv->column > 0 && kv->column < kv->term_width - 26) {
-            kv->column += fprintf(output->file, "%*s", 25 - kv->column % 26, " ");
-        }
-
-        // print key
-        char *key = *data->pretty_key ? data->pretty_key : data->key;
-        kv->column += fprintf(output->file, "%-10s: ", key);
-        // print value
-        if (color)
-            term_set_fg(kv->term, kv_color_for_key(data->key));
-        print_value(output, data->type, data->value, data->format);
-        if (color)
-            term_set_fg(kv->term, TERM_COLOR_RESET);
-
-        // force break after some known keys
-        if (kv->column > 0 && kv_break_after_key(data->key)) {
-            kv->column = kv->term_width; // force break;
-        }
-
-        data = data->next;
-    }
-    --kv->data_recursion;
-
-    // top-level: always end with newline
-    if (!kv->data_recursion && kv->column > 0) {
-        //fprintf(output->file, "\n"); // data_output_print() already adds a newline
-        kv->column = 0;
-    }
-}
-
-static void print_kv_array(data_output_t *output, data_array_t *array, char const *format)
-{
-    //fprintf(output->file, "[ ");
-    for (int c = 0; c < array->num_values; ++c) {
-        if (c)
-            fprintf(output->file, ", ");
-        print_array_value(output, array, format, c);
-    }
-    //fprintf(output->file, " ]");
-}
-
-static void print_kv_double(data_output_t *output, double data, char const *format)
-{
-    data_output_kv_t *kv = (data_output_kv_t *)output;
-
-    kv->column += fprintf(output->file, format ? format : "%.3f", data);
-}
-
-static void print_kv_int(data_output_t *output, int data, char const *format)
-{
-    data_output_kv_t *kv = (data_output_kv_t *)output;
-
-    kv->column += fprintf(output->file, format ? format : "%d", data);
-}
-
-static void print_kv_string(data_output_t *output, const char *data, char const *format)
-{
-    data_output_kv_t *kv = (data_output_kv_t *)output;
-
-    kv->column += fprintf(output->file, format ? format : "%s", data);
-}
-
-static void data_output_kv_free(data_output_t *output)
-{
-    data_output_kv_t *kv = (data_output_kv_t *)output;
-
-    if (!output)
-        return;
-
-    if (kv->color)
-        term_free(kv->term);
-
-    free(output);
-}
-struct data_output *data_output_kv_create(FILE *file)
-{
-    data_output_kv_t *kv = calloc(1, sizeof(data_output_kv_t));
-    if (!kv) {
-        WARN_CALLOC("data_output_kv_create()");
-        return NULL; // NOTE: returns NULL on alloc failure.
-    }
-
-    kv->output.print_data   = print_kv_data;
-    kv->output.print_array  = print_kv_array;
-    kv->output.print_string = print_kv_string;
-    kv->output.print_double = print_kv_double;
-    kv->output.print_int    = print_kv_int;
-    kv->output.output_free  = data_output_kv_free;
-    kv->output.file         = file;
-
-    kv->term = term_init(file);
-    kv->color = term_has_color(kv->term);
-
-    kv->ring_bell = 0; // TODO: enable if requested...
-
-    return &kv->output;
-}
-
-/* CSV printer; doesn't really support recursive data objects yet */
-
-typedef struct {
-    struct data_output output;
-    const char **fields;
-    int data_recursion;
-    const char *separator;
-} data_output_csv_t;
-
-static void print_csv_data(data_output_t *output, data_t *data, char const *format)
-{
-    data_output_csv_t *csv = (data_output_csv_t *)output;
-
-    const char **fields = csv->fields;
-    int i;
-
-    if (csv->data_recursion)
-        return;
-
-    int regular = 0; // skip "states" output
-    for (data_t *d = data; d; d = d->next) {
-        if (!strcmp(d->key, "msg") || !strcmp(d->key, "codes") || !strcmp(d->key, "model")) {
-            regular = 1;
-            break;
-        }
-    }
-    if (!regular)
-        return;
-
-    ++csv->data_recursion;
-    for (i = 0; fields[i]; ++i) {
-        const char *key = fields[i];
-        data_t *found = NULL;
-        if (i)
-            fprintf(output->file, "%s", csv->separator);
-        for (data_t *iter = data; !found && iter; iter = iter->next)
-            if (strcmp(iter->key, key) == 0)
-                found = iter;
-
-        if (found)
-            print_value(output, found->type, found->value, found->format);
-    }
-    --csv->data_recursion;
-}
-
-static void print_csv_array(data_output_t *output, data_array_t *array, char const *format)
-{
-    for (int c = 0; c < array->num_values; ++c) {
-        if (c)
-            fprintf(output->file, ";");
-        print_array_value(output, array, format, c);
-    }
-}
-
-static void print_csv_string(data_output_t *output, const char *str, char const *format)
-{
-    data_output_csv_t *csv = (data_output_csv_t *)output;
-
-    while (*str) {
-        if (strncmp(str, csv->separator, strlen(csv->separator)) == 0)
-            fputc('\\', output->file);
-        fputc(*str, output->file);
-        ++str;
-    }
-}
-
-static int compare_strings(const void *a, const void *b)
-{
-    return strcmp(*(char **)a, *(char **)b);
-}
-
-static void data_output_csv_start(struct data_output *output, const char **fields, int num_fields)
-{
-    data_output_csv_t *csv = (data_output_csv_t *)output;
-
-    int csv_fields = 0;
-    int i, j;
-    const char **allowed = NULL;
-    int *use_count = NULL;
-    int num_unique_fields;
-    if (!csv)
-        goto alloc_error;
-
-    csv->separator = ",";
-
-    allowed = calloc(num_fields, sizeof(const char *));
-    if (!allowed) {
-        WARN_CALLOC("data_output_csv_start()");
-        goto alloc_error;
-    }
-    memcpy(allowed, fields, sizeof(const char *) * num_fields);
-
-    qsort(allowed, num_fields, sizeof(char *), compare_strings);
-
-    // overwrite duplicates
-    i = 0;
-    j = 0;
-    while (j < num_fields) {
-        while (j > 0 && j < num_fields &&
-                strcmp(allowed[j - 1], allowed[j]) == 0)
-            ++j;
-
-        if (j < num_fields) {
-            allowed[i] = allowed[j];
-            ++i;
-            ++j;
-        }
-    }
-    num_unique_fields = i;
-
-    csv->fields = calloc(num_unique_fields + 1, sizeof(const char *));
-    if (!csv->fields) {
-        WARN_CALLOC("data_output_csv_start()");
-        goto alloc_error;
-    }
-
-    use_count = calloc(num_unique_fields + 1, sizeof(*use_count)); // '+ 1' so we never alloc size 0
-    if (!use_count) {
-        WARN_CALLOC("data_output_csv_start()");
-        goto alloc_error;
-    }
-
-    for (i = 0; i < num_fields; ++i) {
-        const char **field = bsearch(&fields[i], allowed, num_unique_fields, sizeof(const char *),
-                compare_strings);
-        int *field_use_count = use_count + (field - allowed);
-        if (field && !*field_use_count) {
-            csv->fields[csv_fields] = fields[i];
-            ++csv_fields;
-            ++*field_use_count;
-        }
-    }
-    csv->fields[csv_fields] = NULL;
-    free(allowed);
-    free(use_count);
-
-    // Output the CSV header
-    for (i = 0; csv->fields[i]; ++i) {
-        fprintf(csv->output.file, "%s%s", i > 0 ? csv->separator : "", csv->fields[i]);
-    }
-    fprintf(csv->output.file, "\n");
-    return;
-
-alloc_error:
-    free(use_count);
-    free(allowed);
-    if (csv)
-        free(csv->fields);
-    free(csv);
-}
-
-static void print_csv_double(data_output_t *output, double data, char const *format)
-{
-    fprintf(output->file, "%.3f", data);
-}
-
-static void print_csv_int(data_output_t *output, int data, char const *format)
-{
-    fprintf(output->file, "%d", data);
-}
-
-static void data_output_csv_free(data_output_t *output)
-{
-    data_output_csv_t *csv = (data_output_csv_t *)output;
-
-    free(csv->fields);
-    free(csv);
-}
-
-struct data_output *data_output_csv_create(FILE *file)
-{
-    data_output_csv_t *csv = calloc(1, sizeof(data_output_csv_t));
-    if (!csv) {
-        WARN_CALLOC("data_output_csv_create()");
-        return NULL; // NOTE: returns NULL on alloc failure.
-    }
-
-    csv->output.print_data   = print_csv_data;
-    csv->output.print_array  = print_csv_array;
-    csv->output.print_string = print_csv_string;
-    csv->output.print_double = print_csv_double;
-    csv->output.print_int    = print_csv_int;
-    csv->output.output_start = data_output_csv_start;
-    csv->output.output_free  = data_output_csv_free;
-    csv->output.file         = file;
-
-    return &csv->output;
-}
-
 /* JSON string printer */
 
 typedef struct {
@@ -1023,156 +528,4 @@ size_t data_print_jsons(data_t *data, char *dst, size_t len)
     format_jsons_object(&jsons.output, data, NULL);
 
     return len - jsons.msg.left;
-}
-
-/* Datagram (UDP) client */
-
-typedef struct {
-    struct sockaddr_storage addr;
-    socklen_t addr_len;
-    SOCKET sock;
-} datagram_client_t;
-
-static int datagram_client_open(datagram_client_t *client, const char *host, const char *port)
-{
-    if (!host || !port)
-        return -1;
-
-    struct addrinfo hints, *res, *res0;
-    int    error;
-    SOCKET sock;
-
-    memset(&hints, 0, sizeof(hints));
-    hints.ai_family = PF_UNSPEC;
-    hints.ai_socktype = SOCK_DGRAM;
-    hints.ai_flags = AI_ADDRCONFIG;
-    error = getaddrinfo(host, port, &hints, &res0);
-    if (error) {
-        fprintf(stderr, "%s\n", gai_strerror(error));
-        return -1;
-    }
-    sock = INVALID_SOCKET;
-    for (res = res0; res; res = res->ai_next) {
-        sock = socket(res->ai_family, res->ai_socktype, res->ai_protocol);
-        if (sock >= 0) {
-            client->sock = sock;
-            memset(&client->addr, 0, sizeof(client->addr));
-            memcpy(&client->addr, res->ai_addr, res->ai_addrlen);
-            client->addr_len = res->ai_addrlen;
-            break; // success
-        }
-    }
-    freeaddrinfo(res0);
-    if (sock == INVALID_SOCKET) {
-        perror("socket");
-        return -1;
-    }
-
-    //int broadcast = 1;
-    //int ret = setsockopt(client->sock, SOL_SOCKET, SO_BROADCAST, &broadcast, sizeof(broadcast));
-
-    return 0;
-}
-
-static void datagram_client_close(datagram_client_t *client)
-{
-    if (!client)
-        return;
-
-    if (client->sock != INVALID_SOCKET) {
-        closesocket(client->sock);
-        client->sock = INVALID_SOCKET;
-    }
-
-#ifdef _WIN32
-    WSACleanup();
-#endif
-}
-
-static void datagram_client_send(datagram_client_t *client, const char *message, size_t message_len)
-{
-    int r =  sendto(client->sock, message, message_len, 0, (struct sockaddr *)&client->addr, client->addr_len);
-    if (r == -1) {
-        perror("sendto");
-    }
-}
-
-/* Syslog UDP printer, RFC 5424 (IETF-syslog protocol) */
-
-typedef struct {
-    struct data_output output;
-    datagram_client_t client;
-    int pri;
-    char hostname[_POSIX_HOST_NAME_MAX + 1];
-} data_output_syslog_t;
-
-static void print_syslog_data(data_output_t *output, data_t *data, char const *format)
-{
-    data_output_syslog_t *syslog = (data_output_syslog_t *)output;
-
-    // we expect a normal message around 500 bytes
-    // full stats report would be 12k and we want a max of MTU anyway
-    char message[1024];
-    abuf_t msg = {0};
-    abuf_init(&msg, message, sizeof(message));
-
-    time_t now;
-    struct tm tm_info;
-    time(&now);
-#ifdef _WIN32
-    gmtime_s(&tm_info, &now);
-#else
-    gmtime_r(&now, &tm_info);
-#endif
-    char timestamp[21];
-    strftime(timestamp, 21, "%Y-%m-%dT%H:%M:%SZ", &tm_info);
-
-    abuf_printf(&msg, "<%d>1 %s %s rtl_433 - - - ", syslog->pri, timestamp, syslog->hostname);
-
-    msg.tail += data_print_jsons(data, msg.tail, msg.left);
-    if (msg.tail >= msg.head + sizeof(message))
-        return; // abort on overflow, we don't actually want to send more than fits the MTU
-
-    size_t abuf_len = msg.tail - msg.head;
-    datagram_client_send(&syslog->client, message, abuf_len);
-}
-
-static void data_output_syslog_free(data_output_t *output)
-{
-    data_output_syslog_t *syslog = (data_output_syslog_t *)output;
-
-    if (!syslog)
-        return;
-
-    datagram_client_close(&syslog->client);
-
-    free(syslog);
-}
-
-struct data_output *data_output_syslog_create(const char *host, const char *port)
-{
-    data_output_syslog_t *syslog = calloc(1, sizeof(data_output_syslog_t));
-    if (!syslog) {
-        WARN_CALLOC("data_output_syslog_create()");
-        return NULL; // NOTE: returns NULL on alloc failure.
-    }
-#ifdef _WIN32
-    WSADATA wsa;
-
-    if (WSAStartup(MAKEWORD(2,2),&wsa) != 0) {
-        perror("WSAStartup()");
-        free(syslog);
-        return NULL;
-    }
-#endif
-
-    syslog->output.print_data   = print_syslog_data;
-    syslog->output.output_free  = data_output_syslog_free;
-    // Severity 5 "Notice", Facility 20 "local use 4"
-    syslog->pri = 20 * 8 + 5;
-    gethostname(syslog->hostname, _POSIX_HOST_NAME_MAX + 1);
-    syslog->hostname[_POSIX_HOST_NAME_MAX] = '\0';
-    datagram_client_open(&syslog->client, host, port);
-
-    return &syslog->output;
 }
